Added simple test that uses flash url loading to cross-domain.
This test is used to demonstrate that cross domain policies provided
over an xmlsocket or in-line with the http protocol (but using the
flash xml protocol) on the same port only authorize flash socket
connections, not http URL loading connections. Therefore the policies
can be different between the two methods of loading, allowing raw
sockets (which do not have access to cross-domain cookies) to be more
permissive than URL loaders (which will send cross-domain cookies).