Correctly detect webpack when trying to polyfill utils.setImmediate. Fixes regression in #385
When using a stock webpack config (which injects `process`), calling
`forge.utils.setImmedate` fails with `invalid calling object` on IE11
and Edge. A fix for this was implemented in #390 but the detection didn't
work properly. By checking also for `process.browser`, we can now properly
detect a webpack / browser environment and use the proper workaround for
the error (wrapping `setImmediate`)