x509: add support for Netscape comment extension
This extension is in theory deprecated, but remains widely used, and
supported by openssl and other tools.  It is a useful method to associate a
comment or memo with a certificate.

OID 2.16.840.1.113730.1.13 requires IA5STRING value for the comment.  We filter
for this requirement, and also to avoid creating pathologically long strings
by default.