Check for NativeBuffer in ed25519#messageToNativeBuffer
While this technically is not necessary, because node.js' Buffer
inherits from Uint8Array this contract is broken by jest.
See facebook/jest#4422.

Add a `message instanceof NativeBuffer` to the function to fix the issue,
it does no harm and *technically* would be the more correct check, because
for node.js passing a raw Uint8Array does not return a Buffer.